Added Living Entites (Hunger, Thirst) and improved Movement (positions as floats), added simple AI

+++ src/geras/jmoon/GameStates/WorldGameState.java
@@ -50,16 +50,16 @@ public class WorldGameState extends BasicTWLGameState {
 		int midY = container.getHeight() / 2;
 		
 		//top left pixel of the map, if it were drawn completely
-		int mapTopX = Math.max(Settings.resolutionX - Settings.mapWidth * Settings.tileWidth, Math.min(0,midX - player.getPosX()));
-		int mapTopY = Math.max(Settings.resolutionY - Settings.mapHeight * Settings.tileHeight, Math.min(0, midY - player.getPosY()));
+		int mapTopX = Math.max(Settings.resolutionX - Settings.mapWidth * Settings.tileWidth, Math.min(0,midX - (int)player.getPosX()));
+		int mapTopY = Math.max(Settings.resolutionY - Settings.mapHeight * Settings.tileHeight, Math.min(0, midY - (int)player.getPosY()));
 		
 		//width and height to be drawn
 		int width = Settings.resolutionX / Settings.tileWidth + 4;
 		int height = Settings.resolutionY / Settings.tileHeight + 4;
 		
 		//first Field to be drawn
-		int topLeftFieldX = Math.min(Settings.mapWidth - width, Math.max(0, player.getPosX() / Settings.tileWidth - (width / 2)));
-		int topLeftFieldY = Math.min(Settings.mapHeight - height, Math.max(0, player.getPosY() / Settings.tileHeight - (height / 2)));
+		int topLeftFieldX = Math.min(Settings.mapWidth - width, Math.max(0, (int)player.getPosX() / Settings.tileWidth - (width / 2)));
+		int topLeftFieldY = Math.min(Settings.mapHeight - height, Math.max(0, (int)player.getPosY() / Settings.tileHeight - (height / 2)));
 		
 		//Position where the first field should be drawn
 		int topLeftDrawX = mapTopX + topLeftFieldX * Settings.tileHeight;
